requirements.txt
numpy
Pillow
scipy
noise
